requirements.txt
Django==1.6.1
South==0.8.4
argparse==1.2.1
diff-match-patch==20110725.1
django-extensions==1.2.5
django-import-export==0.1.5
django-suit==0.2.5
six==1.4.1
tablib==0.9.11
wsgiref==0.1.2
